검색기능 수정 / 데이터가공 - 김다율

+++ src/main/java/com/green/booktodolist/plan/PlanMapper.java
@@ -2,15 +2,15 @@ package com.green.booktodolist.plan;
 
 import com.green.booktodolist.plan.model.PlanBookDataDto;
 import com.green.booktodolist.plan.model.PlanBookDto;
-import com.green.booktodolist.plan.model.PlanBookInsDto;
 import com.green.booktodolist.plan.model.PlanTodoDto;
+import com.green.booktodolist.plan.model.PlanTodoInsDto;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
 
 @Mapper
 public interface PlanMapper {
-    int insBook(PlanBookDto dto);
+    Long insBook(PlanBookDto dto);
     int insTodoList(PlanTodoDto dto);
     List<PlanBookDataDto> callapihttp();
 }
